+++ CHANGELOG.md
@@ -2,7 +2,19 @@
 
 ## Unreleased
 
-## v0.1.3 - 2021-05-19
+## v0.1.4 - 2021-05-11
+
+- Add support for `--prompt-mfa` to force promting for which MFA you want to use
+- Add support for `flush` command
+- Various code clean up & refactoring
+- Fix crash in `list` command when account aliases are not defined
+- Add support for `ONELOGIN_AWS_DURATION` environment variable
+- Improve `version` command
+- Remove all traces of `duration` from config file as it was never supported
+- Change log level argument to `-L`
+- Add support for `--lines` argument 
+
+## v0.1.3 - 2021-05-10
 
 - Fix SecTrustedApplicationCreateFromPath deprecation warning